https://github.com/OpenPaaS-Suite/openpaas-gatling/issues/32: Update Calendar & Contacts simulation
